Fix fbterm-tmux-bash to remove ugly "dummy window"
The problem was:

* fbterm uses the command it's given when starting a new window
* usually, this command needs to attach to the existing tty session
* however, if this tty session doesn't exist, it needs to be created.

The script used to create the session outside the fbterm command;
however, as tmux doesn't support "headless" sessions without windows, it
created the session with a "dummy window" that would be removed later.

This mostly worked, but had an annoying side-effect: since at the time
the first "real" window was created, the dummy window was taking up the
"0" slot, the real window would get the "1" slot. The second real window
would then go into the "0" slot later.

The current solution moves this work into the command that fbterm runs,
at the cost of one additional sh process per fbterm window (which IMO
is the lesser evil).