Add sshadd function to replace ssh-add alias
The alias was getting inconvenient with commands like `ssh-add -l`. I
briefly tried replacing it with a function ssh-add which did the old
behavior (add all keys) if there were no arguments and otherwise just
forwarded to the real ssh-add, but `ssh-add` is not a valid function
name in sh mode, which causes problems with X initialization, so let’s
just have a separate function and leave the real ssh-add alone.