Add support for existing sessions to lterm
If a tmux session with that name already exists, raise the corresponding
window if it exists or create a new window attaching to that session.

Previously, using lterm with an existing tmux session would fail, since
tmux new-window would refuse to create a session with an already taken
name. The user would see no indication of this, since the tmux error
message would be displayed in the terminal window which exits with tmux.

When attaching to the session, we set the working directory, because
apparently tmux supports that. This means that lterm can also be used to
update the working directory of a tmux session (e. g., after the symlink
in the config directory was updated).