Exit shell if lterm is first command
The effect of this is that if you open a new terminal and run `lterm
whatever`, the spawning terminal silently closes itself and you can just
stay in the `whatever` terminal without wasting another thought on the
other one. This makes it even cheaper to raise any terminal with `lterm
whatever`, even if it would also just be a few Alt+Tabs away.