Explictly set activate-window-gnome-shell IFS
I’m not sure why, but I got an error indicating that the output wasn’t
being split correctly. I suppose it’s a good idea to set IFS here
anyways, because we only want to split on spaces, not tabs or newlines
(not that the output should contain any).