Refactor term/lterm/rterm
Extract parts of lterm into a new term function, and rename the old term
function into rterm (remote term), using the new term function. The
practical effect of this is that rterm no longer uses the deprecated
gnome-terminal -e option and also closes the summoning shell if it’s the
first command, just like lterm already did.

I’m not quite happy with the way that the auto-closing code is partially
duplicated between term and lterm, but it will do for now. Perhaps I can
think of something better later.