Fix exit-if-first-command logic of term
If I understand it correctly, BASH_LINENO without any extra array
notation expands to the first element of the array, which is the line
number in the current function, so I’m not sure how this ever worked :)
Instead use the last element, which is the top of the call stack (i. e.
the interactive shell). Also check if the shell is interactive in the
flags, and remove the BASH_SOURCE introspection – if someone calls term
directly without going through lterm or rterm, we might as well apply
the same logic, and this is probably more stable.