Add `man` function that also supports builtins
I understand why the distinction between `help` and `man` exists
(builtins are shell specific but the manpage database is shared between
all shells), but it’s still annoying from a user’s perspective – when I
type `man cd`, I’m usually not interested in the POSIX version of the
command (I would write `man 1p cd` for that), but that’s the only one
that has a manpage. This function resolves that issue by trying `help`
first and falling back to `man` if there’s no `help` page for the topic.