Refresh tmux gnome-terminal environment when possible
The tmux server initializes its environment when it is first started,
and copies that into each new session. Since I have gnome-terminal set
to run `tmux` as command (instead of my login shell), this means that
the tmux server starts out with the gnome-terminal environment variables
of the terminal which launched it.

However, if subsequently all terminal windows are closed while the tmux
server remains alive (typically because at least one session was
detached from and not terminated), this means that the next time I start
a terminal, it will start a new gnome-terminal-server, and the values in
the environment of the tmux server and all tmux sessions will be
outdated. This leads to various errors when trying to launch new
terminals from those sessions (e. g.  via my lterm and rterm functions),
such as:

    # Failed to use specified server: GDBus.Error:org.freedesktop.DBus.Error.ServiceUnknown: The name :1.159 was not provided by any .service files
    # Falling back to default server.

(referring to an old GNOME_TERMINAL_SERVICE; in this case, the terminal
can still be opened), or:

    # Error creating terminal: failed to get screen from object path /org/gnome/Terminal/screen/[UUID]

(referring to an old GNOME_TERMINAL_SCREEN; in this case, the terminal
fails to open).

Mitigate this by updating the variables in the tmux environment each
time a new session is opened or attached, which should mean getting the
variables from the newly created terminal window and therefore obtaining
their current, “fresh” value.

Note that this doesn’t help when trying to launch a new terminal from an
existing shell within one of those tmux sessions, since the shell’s
environment will not be updated automatically; however, you can simply
create a new tmux window to launch the new terminal (recall that both
lterm and rterm will auto-close their auxiliary shell in that case).