Configure notes.rewriteRef
Git defaults to keeping notes when amending a commit, but only if this
setting is also set, and the useful value of refs/notes/commits is only
documented in git-config(1), not in git-notes(1). This seems
unexpectedly user-hostile to me, but whatever, let’s configure it.

(I mainly use git notes to attach additional comments to a patch file
I’m sending out with format-patch or send-email. This also requires
passing the --notes option to those commands, for which there is no
configuration setting at all. Oh well.)