Enable mouse mode in tmux
Like xterm-mouse-mode in Emacs (see 96f38ee2dd), this is also very
useful, as long as you know how to disable it when not needed
(Shift+Mouse). This improved mouse mode was added in tmux 2.1, and I
think all the systems I have access to have a version newer than that.