Add ssh configuration
OpenSSH 7.3 added the Include directive, which allows me to split up the
configuration file and publicly commit some parts of it. Apart from the
main configuration file committed here, I also have one file in
~/.ssh/config.d per host that I usually connect to, starting with a Host
line and then adding other configuration (HostName, IdentityFile, User,
etc.). Much of that could actually be public as well, but unfortunately
I can’t commit those files in their current state because the
IdentityFile contains the local hostname and is therefore host-specific
and not suitable for sharing across systems. But perhaps I’ll improve on
that situation in the future.