Add make-install function
Similarly to the iotop function, this is a more restricted alternative
to simply running a command as sudo. In this case, you replace

    make
    sudo make install

with

    make
    make-install

where make-install still runs `make install` as root, but removes almost
all root-specific privileges from the process (because file system
manipulation is the only reason it should need root, and even that only
in a restricted set of directories).

Note that this only works for makefiles where the install step doesn’t
need to modify anything in the build directory anymore; that’s the
ideal, but it’s not always the case in more complicated projects (for
example, I couldn’t get this to work with the git build process). It
also doesn’t provide an analogous command to run `make uninstall` yet.