Add systemd unit files to back up Toolforge databases
These unit templates can be used for opportunistic, daily backups of
tool databases (when the Toolforge SSH key is not available in the SSH
agent, that day’s backup is skipped). The symlink in
~/.config/user/timers.target.wants/ sets up such a backup for the
QuickCategories tool; others can be added with commands like this:

    systemctl --user enable --now toolsdb-backup@TOOL.DATABASE.timer

Backups are stored in toolsdb-backup/TOOL/DATABASE/ under
~/.local/share/ (or wherever $XDG_DATA_HOME points).

This assumes that the SSH configuration contains a host nickname
“toolforge” for the Toolforge login bastion, with the right settings
(username, possibly identity file, etc.). It also requires the SSH agent
to be available to other systemd user units (in one way or another, the
user manager’s environment should contain the SSH_AUTH_SOCK variable).