Use hash for SSH ControlPath
A ControlPath with the remote username, hostname and port can be too
long for a UNIX domain socket (108 bytes on Linux), since the directory
is already quite long. Instead, use %C, which is the hash of all this
information plus the local hostname, and always has constant length.
This does mean that it’s no longer feasible to delete just one
particular control socket, but in practice I haven’t found that to be a
very limiting restriction so far.