Clone more efficiently from Wikimedia Gerrit
Use Git protocol version 2 and add remote aliases, similar to the
existing gh: family, to make using it easier (Git protocol v2 is
apparently not available over ssh). This cuts the time of a no-op fetch
roughly in half.

wmg: is the general alias for Wikimedia Gerrit; wmge: is a shortcut for
the common case of referring to an extension:

    git clone -o gerrit wmge:ExtensionName(.git)

I updated the remotes of the extensions I already cloned with the
following snippet:

    for file in extensions/*/.git/config; do
        sed -i.backup-2020-06-29-wmge 's|gerrit:mediawiki/extensions/|wmge:|' -- "$file";
    done