Don’t overwrite existing ToolsDB backup
By default, the backup only runs once a day, so the target file
shouldn’t exist yet. If it does exist, either the existing backup or the
new one must be manual, and it’s probably being repeated for a reason,
so instead of just overwriting the data file, we should error out and
make the user handle the situation (either move the existing file out of
the way or, if it’s really not wanted, remove it).