Reimplement activate-window using Gnome Shell extension
The activate-window function, used by lterm when the given tmux session
already exists and may have an open window that should be activated,
broke with Gnome 41, because org.gnome.Shell.Eval is now restricted to
“unsafe mode”. To deal with this, I wrote a small, custom Gnome Shell
extension [1][2] that exposes a D-Bus interface for activating a window
by its title. activate-window now tries to use this interface first.

In the extension, I didn’t want to expose a method to activate a window
whose title matches a given regular expression, because that would make
the user vulnerable to ReDoS attacks; for the lterm use case, I only
want to activate the window with the given prefix anyways. Therefore,
the code in the two fallbacks (Eval, xdotool) is adjusted to also no
longer accept regular expressions and do prefix matching. (In the
xdotool case, the matching is still case-insensitive, there’s nothing we
can do about that.) This is pretty much a shot in the dark, though – I
haven’t tested these fallbacks very much, because I’m not very
interested in them.

[1]: https://extensions.gnome.org/extension/5021/activate-window-by-title/
[2]: https://github.com/lucaswerkmeister/activate-window-by-title