Add .gpgedit command
The idea is similar to sudoedit – temporarily make a file available for
editing, automatically restoring it at the end – but by only supporting
Emacs instead of arbitrary editors, we can handle the decrypted file
contents in a way that makes them mostly inaccessible to the rest of the
system, using stdin/stdout instead of a temporary file.

The recipient ID is hard-coded, because that’s all I need, and making it
more generic is difficult given gpg’s command line interface (there’s no
easy and stable way to get the recipient(s) of an encrypted file, nor to
encrypt with the same settings as a given reference file).