gpgedit: Don’t save empty file
The emacs usage (C-x C-c without C-x C-s) isn’t obvious, so if the user
accidentally exits emacs in a way that leaves the file to appear empty,
don’t save it. If the user wants to empty the file, they can just delete
it, or (if they want to keep an encrypted empty file) freshly encrypt
/dev/null.

Note that, even without this protection, the `mv -b` option leaves
behind a backup of the previous file version, which you can use to
recover from various errors.