toots+tweets: customize bash completion
Only complete files with the right file name extensions. I keep my
archives (of both) in ~/Downloads/, which also has tons of other files,
and if I don’t remember the file name pattern, it can be annoying to
find the right file; this should help, by presenting fewer candidates in
the completion.