Update .emacs
Many of these changes, especially the package list, are years old, and I
just haven’t had the courage / motivation to track them into Git for a
while now (package management and autoload / auto-mode-alist is
confusing). But the vue3-mode parts of this at least make Vue files
work, which is pretty nice.