Add git branchd command to delete rebased branches
This is meant to replace my previous workflow of:

    $ git rebase -i master BRANCH # check that it says noop
    $ git switch master # (see T377562 for "master" name)
    $ git branch -d BRANCH # now deletes cleanly

I previously had a `rad` alias (“rebase and delete”) for this in one
repository:

    !rad() { git rebase main "$1" && git switch main && git branch -d "$1"; }; rad

But since it has to hard-code the name of the “main” branch (which in
most Wikimedia repositories is still called “master”), I never added it
to my general Git config. This `git cherry`-based implementation is
nicer – it doesn’t have the “safeguard” of only using `git branch -d`
(without --force), but that should be okay, given that the logic used by
`git cherry` is the same as that used in the rebase step. (Also, the
`git branch -d/-D` command will still print the branch name and hash
that was just deleted.)